# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DATN E-Commerce API",
        "description": "Tài liệu API hệ thống DATN E-Commerce Backend",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "DATN API Server"
        }
    ],
    "paths": {
...
        "/api/ratings": {
            "post": {
                "tags": [
                    "Đánh giá sản phẩm (Rating)"
                ],
                "summary": "Gửi đánh giá sản phẩm",
                "operationId": "e876bdaebd2a715619d9d19e8be09398",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_item_id",
                                    "rating"
                                ],
                                "properties": {
                                    "order_item_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Giày đi rất êm chân, giao hàng nhanh!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gửi đánh giá thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reviews": {
            "get": {
                "tags": [
                    "Đánh giá sản phẩm (Rating)"
                ],
                "summary": "[Admin] Danh sách đánh giá của khách hàng",
                "operationId": "7555a41bac6f952ebe379bb7e73550d1",
                "parameters": [
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Số sao (1-5 hoặc all)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Trạng thái (pending/replied/hidden/all)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Tìm theo tên/email/nội dung",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reviews/{id}/reply": {
            "post": {
                "tags": [
                    "Đánh giá sản phẩm (Rating)"
                ],
                "summary": "[Admin] Phản hồi đánh giá",
                "operationId": "09103b1a165e17ed218a3a70d7fef906",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reply"
                                ],
                                "properties": {
                                    "reply": {
                                        "type": "string",
                                        "example": "Cảm ơn bạn đã tin tưởng SaigonShoes!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phản hồi thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reviews/{id}/status": {
            "put": {
                "tags": [
                    "Đánh giá sản phẩm (Rating)"
                ],
                "summary": "[Admin] Cập nhật trạng thái hiển thị/ẩn đánh giá",
                "operationId": "3edce7517aba4767c069d9eadd5ac7d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "replied",
                                            "hidden"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reviews/{id}": {
            "delete": {
                "tags": [
                    "Đánh giá sản phẩm (Rating)"
                ],
                "summary": "[Admin] Xóa đánh giá",
                "operationId": "0364246127ac99083be5ac8451bcf210",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập token của bạn dạng: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}